Update for dbus-sharp 0.8 & dbus-sharp-glib 0.6
We now require both dbus-sharp 0.8 and dbus-sharp-glib 0.6.

We were using an obsolete Bus.Session.Register which required a bus
name, and it's now gone. This allows us to remove some
DBusServiceManager.RegisterObject overloads that become unnecessary.

As an additional consequence, we don't need the DBusExportable attribute
anymore. It was used by classes implementing IDBusExportable to indicate
their service name, which was then used for the bus name.