DapService: Correct source disposal
After BGO#666696, a Flush call was added to DapSource Disposal to ensure
that playlists were synchronised when the user quits Banshee while a DAP
is plugged in.

Unfortunately that solution is no longer working, the culprit apparently
is the threading used in DapService during the UnmapDevice call. Since
that method spawns a thread which locks on DapService's sync variable,
all the unmap jobs wait until DapService is disposed, nullifying the
source map, required during the Unmap call, as a consequence, BGO#666696
has re-occurred as the unmap call returns silently.

The solution therefore, is to call the sequential Unmap method which
ensures that sources are disposed before the map is eliminated.