mass-storage: ThreadAssist for Art (bgo#734258)
In addition to wrapping Gdk.Pixbuf handling in a main thread call we
take the opportunity to re-enable the 'shortcut' when a user's device is
expecting jpg cover art.

We also streamline the function a little by taking advantage of the fact
that calling LookupPixbuf on ArtworkManager is identical to calling
LookupScalePixbuf with a size parameter of 0.