mtp: Handle awkward filenames
This commit handles awkward path components when creating parent folders
by escaping certain characters. In addition, we notice that MTP appears
to be case insensitive and add a fix in folder selection to handle this.
Finally, we modify the MakeAlbumKey method to receive a TrackInfo after
noticing it was used with different track properties, nullifying it's
device folder structure caching characteristics.