perf: Don't reload everything
This commit prevents reloading when there is no trigger filter, which
seems to only be he case when filters are added through the View menu.
This seems to have no effect on the active filter behaviour.

The reload is expensive and appears to work in a recursive fashion
leading to a considerable slow down when performed in the course of
normal operation.