dap: PotentialSource is first class
This commit adds PotentialSource as a first class citizen and extends it
from the existing MtpSource. This fits in nicely instead of having the
source UI swap out in the tree-view. It also improves the user
experience by allowing Banshee to recognise a source it could use but
where the OS or player may wish to do other things with it while it is
running intead.