ui: Refactor view configuration
Since PotentialSource was made first class, we introduced a bug where
the source view wouldn't update when the source was connected, still
showing the Claim screen. We move the view configuration to a method
called Widgets and call that from both SourceUpdated and
ActiveSourceUpdated, but only act when the source is the active one,
duplcating some logic.

The classic interface exposes this bug too, we don't deal with that
here, we'll need a base class extraction and merge to do that.

In addition, we remove the commented out track info display lines, in
line with our no-duplication stance on elements, the user can Ctrl-J to
the playing track to see its information.